Deploy SAA as standalone test

Deploy SAA as standalone test

SAA_standalone(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  num_examples = list(long_tones = 2L, arrhythmic = 2L, rhythmic = 0L),
  arrhythmic_item_bank = SAA::Berkowitz_item_bank_subset,
  rhythmic_item_bank = SAA::Berkowitz_item_bank_subset,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  examples = 2L,
  final_results = TRUE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  melody_length = c(4, 20),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = NULL,
  concise_wording = TRUE,
  skip_setup = FALSE,
  full_screen = FALSE,
  additional_scoring_measures = NULL,
  default_range = NULL,
  long_tone_paradigm = c("simultaneous_recall", "call_and_response"),
  get_p_id = FALSE,
  languages = c("en", "de", "it", "lv"),
  volume_meter_on_melody_trials = FALSE,
  volume_meter_on_melody_trials_type = "default",
  long_tone_length = 5,
  allow_SNR_failure = FALSE,
  requirements_page = TRUE,
  report_SNR = FALSE,
  show_introduction = TRUE,
  show_instructions = TRUE,
  asynchronous_api_mode = FALSE,
  experiment_id = NULL,
  user_id = NULL,
  get_answer_melodic = musicassessr::get_answer_pyin_melodic_production,
  content_border = "1px solid #e8e8e8",
  css = system.file("www/css/musicassessr.css", package = "musicassessr"),
  sample_item_bank_via_api = FALSE,
  pass_items_through_url_parameter = FALSE,
  show_intro_text = TRUE,
  show_microphone_type_page = TRUE,
  num_items_review = list(long_tones = 0L, arrhythmic = 0L, rhythmic = 0L),
  logo_url = NULL,
  ...
)

Arguments

app_name

Name of app.

num_items

The number of items as a list.

num_examples

The number of example items as a list.

arrhythmic_item_bank

The item bank (created with itembankr) to deployed with the test.

rhythmic_item_bank

The item bank (created with itembankr) to deployed with the test.

demographics

Deploy demographic form?

demo

Is demo?

feedback

Give feedback after trials?

admin_password

psychTestR admin password.

SNR_test

Deploy signal-to-noise ratio test?

get_range

Deploy a test to get the users range at test time and present stimuli accordingly?

absolute_url

If using online, absolute URL?

examples

No of examples.

final_results

Display final results?

test_username

Is there a username for the user? This is different from a p_id.

gold_msi

Deploy Gold-MSI form?

with_final_page

Should there be a final page? FALSE if there will be more pages in the timeline.

melody_length

What melody lengths should the test be constrained to?

melody_sound

Sound of melody? e.g, piano.

adjust_range

Should the range of the user, recorded at test time, be adjusted based on heuristics?

test_name

Custom name of the test.

show_socials

Should social media sharing options be shown at the end?

headphones_test

Should there be a headphone test?

get_user_info

Grab user info via the browser?

microphone_test

Deploy a microphone test?

allow_repeat_SNR_tests

Logical. TRUE if participant can fail the SNR test threshold and try again.

append_trial_block_before

A list of pages to go before the test.

append_trial_block_after

A list of pages to go after the test.

stop_recording_after

Stop recording after a certain amount of time.

max_goes

How many goes can the user have per melody?

max_goes_forced

Is this forced or optional?

long_tone_trials_as_screening

Should long tone trials be used as a screening mechanism?

long_tone_trials_as_screening_failure_page

Where should users be directed to if they fail the long tone screening?

success_on_completion_page

Where should users be directed to when they complete successfully?

concise_wording

TRUE for more detailed (but longer) instructions.

skip_setup

TRUE to skip setup steps.

full_screen

Should app be full screen?

additional_scoring_measures

A function or list of functions with additional measures for scoring pYIN data.

default_range

A list of the range that stimuli should be presented in, if not collected at test time.

long_tone_paradigm

Can be simultaneous_recall or call_and_response.

get_p_id

if TRUE, get the participant to enter their ID at the beginning of the test.

languages

The available languages.

volume_meter_on_melody_trials

Should there be a volume meter displayed on melody trial pages?

volume_meter_on_melody_trials_type

If so, what type? Can be 'default' or 'playful'.

long_tone_length

How long should the long tone be in seconds? Default: 5

allow_SNR_failure

If TRUE, allow a participant/researcher to proceed, even if the SNR test is failed.

requirements_page

Show a requirements page?

report_SNR

Report SNR after test?

show_introduction

Should introduction be shown (or skipped)?

show_instructions

Should instructions be shown (or skipped)?

asynchronous_api_mode

Should asynchronous API mode br shown?

experiment_id

Manually give an experiment ID when using musicassessrdb.

user_id

Manually give a user ID when using musicassessrdb.

get_answer_melodic

The get_answer function for melodic trials.

content_border

The psychTestR border style.

css

Path to css stylesheet(s).

sample_item_bank_via_api

Is the item bank being sampled via an API?

pass_items_through_url_parameter

Are items being passed through a URL parameter?

show_intro_text

Should intro text be shown?

show_microphone_type_page

Should you ask the participant what kind of microphone they are using?

num_items_review

Number of review items.

logo_url

A URL for the psychTestR logo image.

...