Deploy the adaptive SAA (aSAA)

Deploy the adaptive SAA (aSAA)

aSAA(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = Berkowitz::Berkowitz_IRT_arrhythmic_scaled,
  irt_model = Berkowitz::lm2.2_scaled,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = "",
  examples = 2,
  final_results = TRUE,
  test_username = "",
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = "",
  concise_wording = TRUE,
  skip_setup = FALSE
)

Arguments

item_bank

The (IRT) item bank that will be used for selecting items.

irt_model

The IRT model that will be used for selecting items.

skip_setup