Deploy adaptive SAA (aSAA) as standalone test

Deploy adaptive SAA (aSAA) as standalone test

aSAA_standalone(
  app_name,
  num_items = list(long_tones = 6L, arrhythmic = 15L, rhythmic = 10L),
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  examples = 2L,
  final_results = TRUE,
  musicassessr_aws = FALSE,
  store_results_in_db = FALSE,
  test_username = character(),
  gold_msi = TRUE,
  with_final_page = TRUE,
  item_length = c(3, 15),
  melody_sound = "piano",
  adjust_range = TRUE,
  test_name = "Singing Ability Assessment",
  show_socials = FALSE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  max_goes = 3L,
  max_goes_forced = FALSE,
  long_tone_trials_as_screening = FALSE,
  long_tone_trials_as_screening_failure_page = "http://www.google.com",
  success_on_completion_page = character(),
  concise_wording = TRUE,
  skip_setup = FALSE,
  ...
)

Arguments

app_name

Name of app.

num_items

The number of items as a list.

demographics

Deploy demographic form?

demo

Is demo?

feedback

Give feedback after trials?

admin_password

psychTestR admin password.

SNR_test

Deploy signal-to-noise ratio test?

get_range

Deploy a test to get the users range at test time and present stimuli accordingly?

absolute_url

If using online, absolute URL?

examples

No of examples.

final_results

Display final results?

musicassessr_aws

Is this being deployed on AWS via the musicassessr setup?

store_results_in_db

Store results in a database?

test_username

Is there a username for the user? This is different from a p_id.

gold_msi

Deploy Gold-MSI form?

with_final_page

Should there be a final page? FALSE if there will be more pages in the timeline.

item_length

What melody lengths should the test be constrained to?

melody_sound

Sound of melody? e.g, piano.

adjust_range

Should the range of the user, recorded at test time, be adjusted based on heuristics?

test_name

Custom name of the test.

show_socials

Should social media sharing options be shown at the end?

headphones_test

Should there be a headphone test?

get_user_info

Grab user info via the browser?

microphone_test

Deploy a microphone test?

allow_repeat_SNR_tests

Logical. TRUE if participant can fail the SNR test threshold and try again.

append_trial_block_before

A list of pages to go before the test.

append_trial_block_after

A list of pages to go after the test.

stop_recording_after

Stop recording after a certain amount of time.

max_goes

How many goes can the user have per melody?

max_goes_forced

Is this forced or optional?

long_tone_trials_as_screening

Should long tone trials be used as a screening mechanism?

long_tone_trials_as_screening_failure_page

Where should users be directed to if they fail the long tone screening?

success_on_completion_page

Where should users be directed to when they complete successfully?

concise_wording

TRUE for more detailed (but longer) instructions.

skip_setup

TRUE to skip setup steps.

...

Value